cloud-functions/ecommerce-sync/requirements.txt
google-cloud-bigquery==3.11.4
requests==2.31.0